cars_project/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.3
django-bootstrap4==22.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.35
Pillow==9.1.0
psycopg2==2.9.3
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
